Author Form Input for Rooms:
I added a window that pops up when you click on a room. Does not
currently work with transparent walls.

Also Tried to move logic out of MapPanel since it is an HIC component. I
made a CIV class which will need to be adjusted to fit MVP but I just
thought I'd use it like this for now.